oidc: Don't fail if server is not reachable
This fixes an issue where LXD would fail on start if the OIDC server
isn't reachable.

Instead of failing on start, `NewVerifier` always returns a Verifier
struct, regardless of whether the OIDC server is reachable or not. If
unreachable, the verifier will try contacting the OIDC server during
the actual authentication. If that fails, so does the authentication.

This also removes the `keySet` variable from the `Verifier` struct as
it's not used anywhere.

Signed-off-by: Thomas Hipp <thomas.hipp@canonical.com>