actions: make linkcheck fail
Sometimes a failing linkcheck can be ignored, but we still want
to see that it is failing.
It's currently the last check, so all other checks will be run
before it fails. If we add checks after this one, they can be
added with condition "always()" to run even if the linkcheck
fails.

Signed-off-by: Ruth Fuchss <ruth.fuchss@canonical.com>