lxd/devices: adjust network device hotplug check
When new net device appears on the host special event
is generated and put in chNetlinkNetwork channel.

We don't want to skip the entire hotplug logic if
netprio cgroup is not present. First of all we have
appropriate check for netprio cgroup in deviceNetworkPriority()
and we should not skip networkAutoAttach.

Instead, let's filter out device removal events.

Signed-off-by: Alexander Mikhalitsyn <aleksandr.mikhalitsyn@canonical.com>