loki: Fix labels names with hyphens
Label names in Loki may not contain any hyphens.

Previously, when setting `loki.labels` to `[context-foo]`, the label
name would be set to `context-foo` as well. The entry however would
never be seen in Loki as it was classified as malformend.

With this change, hyphens are replaced with underscores. In the above
example, the label would now be named `context_foo`.

Signed-off-by: Thomas Hipp <thomas.hipp@canonical.com>