lxd-agent: Fixes vsock listener restart on boot due to vsock module not being fully initialised
Sometimes we were seeing the initial vsock context ID of 4294967295 (equivalent to VMADDR_CID_ANY)
and then a few seconds later the context ID was returning the valid context ID for the VM.

This was previously causing the vsock listener to be restarted 30s after the VM booted, which if
there were any exec sessions ongoing would cause them to be terminated.

This fixes the issue by waiting for a valid vsock ID and ignoring the value 4294967295.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>